More SCC!

+++ arangod/Pregel/Algos/VertexAccumulators/VertexAccumulators.cpp
@@ -197,7 +197,7 @@ struct VertexAccumulatorsMasterContext : MasterContext {
 
       LOG_DEVEL << "no on halt program for this phase, going to next phase " << phase_index;
       aggregate<uint32_t>("phase", phase_index);
-      aggregate<uint32_t>("phase-first-step", globalSuperstep()+1);
+      aggregate<uint64_t>("phase-first-step", globalSuperstep()+1);
       return ContinuationResult::ACTIVATE_ALL;
     }
     // the phase is done, evalute the onHalt program for this phase
